fix: update imports for moved schemas
