+++ src/student_success_tool/dataio/__init__.py
@@ -1,4 +1,4 @@
-from . import pdp
+from . import pdp, schemas
 from .assets import read_config, read_features_table
 from .models import load_mlflow_model, patch_mlflow_model_env_files
 from .read import from_csv_file, from_delta_table
